# docs/swagger.yaml
definitions:
  dto.APICreateShortURLBatchRequestEntry:
    properties:
      correlation_id:
        type: string
      original_url:
        type: string
    type: object
  dto.APICreateShortURLBatchResponseEntry:
    properties:
      correlation_id:
        type: string
      short_url:
        type: string
    type: object
  dto.APICreateShortURLRequest:
    properties:
      url:
        type: string
    type: object
  dto.APICreateShortURLResponse:
    properties:
      error_description:
        type: string
      error_status:
        type: string
      result:
        type: string
    type: object
  dto.APIGetAllURLByUserIDResponseEntry:
    properties:
      original_url:
        type: string
      short_url:
        type: string
    type: object
info:
  contact: {}
  description: Сервис сокращения ссылок
  title: Shortener API
  version: "1.0"
paths:
  /:
    post:
      parameters:
      - description: ссылка которую требуется сократить
        in: body
        name: body
        required: true
        schema:
          type: string
      produces:
      - text/plain
      responses:
        "201":
          description: Created
          schema:
            type: string
        "409":
          description: короткая ссылка уже существует
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервиса
          schema:
            type: string
      summary: Создание короткой ссылки
  /{shortURI}:
    get:
      parameters:
      - description: идентификатор короткой ссылки
        in: path
        name: shortURI
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "307":
          description: Temporary Redirect
          schema:
            type: string
        "404":
          description: короткая ссылка не найдена
          schema:
            type: string
        "410":
          description: короткая ссылка удалена
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервиса
          schema:
            type: string
      summary: получить короткую ссылку
  /api/shorten:
    post:
      parameters:
      - description: запрос на создание короткой ссылки
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.APICreateShortURLRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.APICreateShortURLResponse'
        "400":
          description: ошибка в формате запроса
          schema:
            $ref: '#/definitions/dto.APICreateShortURLResponse'
        "409":
          description: короткая ссылка уже существует
          schema:
            $ref: '#/definitions/dto.APICreateShortURLResponse'
        "500":
          description: внутренняя ошибка сервиса
          schema:
            $ref: '#/definitions/dto.APICreateShortURLResponse'
      summary: Создание короткой ссылки
  /api/shorten/batch:
    post:
      parameters:
      - description: запрос на создание коротких ссылок пачкой
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.APICreateShortURLBatchRequestEntry'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.APICreateShortURLBatchResponseEntry'
            type: array
        "400":
          description: ошибка в формате запроса
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервиса
          schema:
            type: string
      summary: Создание коротких ссылок пачкой
  /api/user/urls:
    delete:
      parameters:
      - description: список идентификаторов коротких ссылок
        in: body
        name: body
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: ошибка в формате запроса
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервиса
          schema:
            type: string
      summary: Удаление коротких ссылок
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.APIGetAllURLByUserIDResponseEntry'
            type: array
        "204":
          description: у пользователя нет коротких ссылок
          schema:
            type: string
        "400":
          description: ошибка в формате запроса
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервиса
          schema:
            type: string
      summary: Получение коротких ссылок созданных пользователем
  /ping:
    get:
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "500":
          description: внутренняя ошибка сервиса
          schema:
            type: string
      summary: получить короткую ссылку
swagger: "2.0"